refactor: repository を completion/ サブディレクトリに整理
service, usecase, valueobject と同様に、repository も
completion/ サブディレクトリに移動し、構造を統一。

- repository/chat.py → repository/completion/chat.py
- repository/completion/__init__.py を追加（カラ）
- 関連するすべてのインポート文を更新

既存のテストがすべてパス (12 tests OK)
