+++ llm_chat/domain/service/completion/chat.py
@@ -11,7 +11,7 @@ from lib.llm.valueobject.config import (
     GeminiConfig,
     ModelName,
 )
-from llm_chat.domain.repository.chat import ChatLogRepository
+from llm_chat.domain.repository.completion.chat import ChatLogRepository
 from llm_chat.domain.service.completion.base import BaseChatService
 from llm_chat.domain.valueobject.completion.message import MessageDTO
 from llm_chat.domain.valueobject.completion.riddle import Gender
