+++ llm_chat/domain/service/completion/rag.py
@@ -5,7 +5,7 @@ from config.settings import BASE_DIR
 from lib.llm.service.completion import OpenAILlmRagService
 from lib.llm.valueobject.config import OpenAIGptConfig, ModelName
 from lib.llm.valueobject.rag import PdfDataloader
-from llm_chat.domain.repository.chat import ChatLogRepository
+from llm_chat.domain.repository.completion.chat import ChatLogRepository
 from llm_chat.domain.service.completion.base import BaseChatService
 from llm_chat.domain.valueobject.completion.message import MessageDTO
