+++ llm_chat/domain/service/completion/riddle.py
@@ -5,7 +5,7 @@ from django.contrib.auth.models import User
 from lib.llm.service.completion import LlmCompletionService, BaseLLMTask
 from lib.llm.valueobject.completion import RoleType, Message
 from lib.llm.valueobject.config import OpenAIGptConfig, GeminiConfig
-from llm_chat.domain.repository.chat import ChatLogRepository
+from llm_chat.domain.repository.completion.chat import ChatLogRepository
 from llm_chat.domain.valueobject.completion.message import MessageDTO
 from llm_chat.domain.valueobject.completion.riddle import (
     Gender,
