+++ llm_chat/domain/usecase/completion/base.py
@@ -4,7 +4,7 @@ from typing import Generator
 from django.contrib.auth.models import User
 
 from lib.llm.valueobject.completion import RoleType, StreamResponse
-from llm_chat.domain.repository.chat import ChatLogRepository
+from llm_chat.domain.repository.completion.chat import ChatLogRepository
 from llm_chat.domain.valueobject.completion.message import MessageDTO
